Add sidebar form and form funcs
Added sidebar elements to the html to hold the form for adding a book.
Added styling to scss to make the sidebar look good and to make the form
more visually appealing. Added func that unhides the form when
the 'new book' button is clicked. Added form funcs that check if a book
is already in the library and then displays the book if does not already
exist there. Added form func that allows the user to insert a new book
to the library