Add func to change read status
Added logic to add another button to the card that calls a function when
clicked. Added said function that prompts the user to input a new status
and then displays that new read status on the card. Refactored original
display books func so that the text was separated from the buttons
semantically and could be targeted better